moving tools/data_scripts to droidlet/tools/data_scripts
fix

fix
