+++ agents/locobot/locobot_agent.py
@@ -12,7 +12,7 @@ from multiprocessing import set_start_method
 import shutil
 
 from droidlet import dashboard
-from tools.data_scripts.try_download import try_download_artifacts
+from droidlet.tools.data_scripts.try_download import try_download_artifacts
 
 if __name__ == "__main__":
     # this line has to go before any imports that contain @sio.on functions
