+++ droidlet/tools/data_scripts/checksum_fn.sh
