+++ droidlet/tools/data_scripts/compute_checksum.sh
