+++ droidlet/tools/data_scripts/default_checksums/craftassist_perception.txt
