+++ droidlet/tools/data_scripts/default_checksums/locobot_perception.txt
