+++ droidlet/tools/data_scripts/fetch_artifacts_from_aws.py
@@ -5,7 +5,7 @@ from subprocess import Popen, PIPE
 import shutil
 
 
-ROOTDIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), '../../')
+ROOTDIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), '../../../')
 print("Rootdir : %r" % ROOTDIR)
