+++ droidlet/tools/data_scripts/fetch_aws_datasets.sh
