+++ droidlet/tools/data_scripts/fetch_aws_models.sh
