+++ droidlet/tools/data_scripts/fetch_internal_resources.py
