+++ droidlet/tools/data_scripts/try_download.sh
