+++ droidlet/tools/data_scripts/upload_datasets_to_aws.sh
