+++ droidlet/tools/data_scripts/upload_models_to_aws.sh
